move abilities to their advanced disc file
