+++ code/code/disc/disc_basic_adventuring.cc
@@ -4,17 +4,12 @@
 #include "handler.h"
 #include "being.h"
 #include "low.h"
-#include "disease.h"
 #include "combat.h"
 #include "disc_basic_adventuring.h"
 #include "disease.h"
 #include "combat.h"
-#include "obj_component.h"
 #include "obj_base_corpse.h"
 #include "obj_tool.h"
-#include "obj_portal.h"
-#include "obj_drinkcon.h"
-#include "pathfinder.h"
 
 void TThing::findBandage(int*) {}
