+++ code/code/disc/disc_basic_combat.cc
@@ -1,9 +1,3 @@
-//////////////////////////////////////////////////////////////////////////
-//
-// SneezyMUD - All rights reserved, SneezyMUD Coding Team
-//
-//////////////////////////////////////////////////////////////////////////
-
 #include "handler.h"
 #include "extern.h"
 #include "being.h"
