+++ code/code/disc/disc_deikhan_mounted.h
@@ -17,3 +17,5 @@ class CDMounted : public CDiscipline {
 
   private:
 };
+
+extern void startChargeTask(TBeing*, const char*);
