+++ code/code/disc/disc_monk_leverage.h
@@ -26,3 +26,5 @@ int shoulderThrow(TBeing*, TBeing*);
 int hurl(TBeing*, TBeing*, char*);
 int defenestrate(TBeing*, TBeing*, sstring);
 int bonebreak(TBeing*, TBeing*);
+int springleap(TBeing*, TBeing*, bool);
+
