+++ code/code/disc/disc_shaman_armadillo.cc
@@ -3,6 +3,8 @@
 #include "room.h"
 #include "being.h"
 #include "disease.h"
+#include "monster.h"
+#include "person.h"
 #include "combat.h"
 #include "spelltask.h"
 #include "disc_shaman_armadillo.h"
@@ -493,4 +495,791 @@ int castCelerite(TBeing* caster, TBeing* victim) {
   } else {
   }
   return TRUE;
+}
+
+
+void shieldOfMists(TBeing* caster, TBeing* victim, TMagicItem* obj) {
+  shieldOfMists(caster, victim, obj->getMagicLevel(),
+    obj->getMagicLearnedness());
+}
+
+int shieldOfMists(TBeing* caster, TBeing* victim) {
+  taskDiffT diff;
+
+  if (!bPassShamanChecks(caster, SPELL_SHIELD_OF_MISTS, victim))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_SHIELD_OF_MISTS]->lag;
+  diff = discArray[SPELL_SHIELD_OF_MISTS]->task;
+
+  start_cast(caster, victim, NULL, caster->roomp, SPELL_SHIELD_OF_MISTS, diff,
+    1, "", rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+int castShieldOfMists(TBeing* caster, TBeing* victim) {
+  int ret, level;
+
+  level = caster->getSkillLevel(SPELL_SHIELD_OF_MISTS);
+  int bKnown = caster->getSkillValue(SPELL_SHIELD_OF_MISTS);
+
+  if ((ret = shieldOfMists(caster, victim, level, bKnown)) == SPELL_SUCCESS) {}
+  return TRUE;
+}
+
+int cheval(TBeing* caster, TBeing* victim, int level, short bKnown) {
+  affectedData aff;
+
+  caster->reconcileHelp(victim, discArray[SPELL_CHEVAL]->alignMod);
+
+  if (caster->bSuccess(bKnown, SPELL_CHEVAL)) {
+    aff.type = SPELL_CHEVAL;
+    aff.level = level;
+    aff.duration = caster->durationModify(SPELL_CHEVAL,
+      (aff.level / 3) * Pulse::UPDATES_PER_MUDHOUR);
+    aff.location = APPLY_NONE;
+    aff.bitvector = 0;
+    switch (critSuccess(caster, SPELL_CHEVAL)) {
+      case CRIT_S_DOUBLE:
+      case CRIT_S_TRIPLE:
+      case CRIT_S_KILL:
+        CS(SPELL_CHEVAL);
+        aff.duration *= 2;
+        break;
+      case CRIT_S_NONE:
+        break;
+    }
+
+    if (!victim->affectJoin(caster, &aff, AVG_DUR_NO, AVG_EFF_YES)) {
+      caster->nothingHappens();
+      return SPELL_FALSE;
+    }
+
+    act("$N seems as if $E is possessed!", FALSE, caster, NULL, victim,
+      TO_NOTVICT, ANSI_GREEN);
+    act("The power of the loa takes dominion inside of you!", FALSE, victim,
+      NULL, NULL, TO_CHAR, ANSI_GREEN);
+    return SPELL_SUCCESS;
+  } else {
+    act("The loa ignore your unfaithful request!", FALSE, caster, NULL, victim,
+      TO_CHAR, ANSI_GREEN);
+    caster->addToLifeforce(-10);
+    caster->nothingHappens(SILENT_YES);
+    return SPELL_FAIL;
+  }
+}
+
+int cheval(TBeing* caster, TBeing* victim) {
+  taskDiffT diff;
+
+  if (!bPassShamanChecks(caster, SPELL_CHEVAL, victim))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_CHEVAL]->lag;
+  diff = discArray[SPELL_CHEVAL]->task;
+
+  start_cast(caster, victim, NULL, caster->roomp, SPELL_CHEVAL, diff, 1, "",
+    rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+void cheval(TBeing* caster, TBeing* victim, TMagicItem* obj) {
+  cheval(caster, victim, obj->getMagicLevel(), obj->getMagicLearnedness());
+}
+
+int castCheval(TBeing* caster, TBeing* victim) {
+  int ret, level;
+
+  level = caster->getSkillLevel(SPELL_CHEVAL);
+  int bKnown = caster->getSkillValue(SPELL_CHEVAL);
+
+  if ((ret = cheval(caster, victim, level, bKnown)) == SPELL_SUCCESS) {
+  } else {
+  }
+  return TRUE;
+}
+
+
+int senseLifeShaman(TBeing* caster, TBeing* victim, int level, short bKnown) {
+  affectedData aff;
+
+  caster->reconcileHelp(victim, discArray[SPELL_SENSE_LIFE_SHAMAN]->alignMod);
+
+  if (caster->bSuccess(bKnown, SPELL_SENSE_LIFE_SHAMAN)) {
+    aff.type = SPELL_SENSE_LIFE_SHAMAN;
+    aff.duration = caster->durationModify(SPELL_SENSE_LIFE_SHAMAN,
+      level * 2 / 3 * Pulse::UPDATES_PER_MUDHOUR);
+    aff.modifier = 0;
+    aff.location = APPLY_NONE;
+    aff.bitvector = AFF_SENSE_LIFE;
+
+    switch (critSuccess(caster, SPELL_SENSE_LIFE_SHAMAN)) {
+      case CRIT_S_DOUBLE:
+      case CRIT_S_TRIPLE:
+      case CRIT_S_KILL:
+        CS(SPELL_SENSE_LIFE_SHAMAN);
+        aff.duration *= 2;
+        break;
+      case CRIT_S_NONE:
+        break;
+    }
+    if (!victim->affectJoin(caster, &aff, AVG_DUR_NO, AVG_EFF_YES)) {
+      caster->nothingHappens();
+      return SPELL_FALSE;
+    }
+    return SPELL_SUCCESS;
+  } else {
+    return SPELL_FAIL;
+  }
+}
+
+void senseLifeShaman(TBeing* caster, TBeing* victim, TMagicItem* obj) {
+  int ret;
+
+  ret = senseLifeShaman(caster, victim, obj->getMagicLevel(),
+    obj->getMagicLearnedness());
+  if (ret == SPELL_SUCCESS) {
+    victim->sendTo("You feel more aware of the world about you.\n\r");
+    act("$n's eyes flicker a faint pale blue.", FALSE, victim, NULL, NULL,
+      TO_ROOM, ANSI_CYAN);
+  } else {
+    caster->nothingHappens();
+  }
+}
+
+int senseLifeShaman(TBeing* caster, TBeing* victim) {
+  taskDiffT diff;
+
+  if (!bPassShamanChecks(caster, SPELL_SENSE_LIFE_SHAMAN, victim))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_SENSE_LIFE_SHAMAN]->lag;
+  diff = discArray[SPELL_SENSE_LIFE_SHAMAN]->task;
+
+  start_cast(caster, victim, NULL, caster->roomp, SPELL_SENSE_LIFE_SHAMAN, diff,
+    1, "", rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+int castSenseLifeShaman(TBeing* caster, TBeing* victim) {
+  int ret, level;
+
+  level = caster->getSkillLevel(SPELL_SENSE_LIFE_SHAMAN);
+  int bKnown = caster->getSkillValue(SPELL_SENSE_LIFE_SHAMAN);
+
+  ret = senseLifeShaman(caster, victim, level, bKnown);
+  if (ret == SPELL_SUCCESS) {
+    victim->sendTo("You feel more aware of the world about you.\n\r");
+    act("$n's eyes flicker a faint pale blue.", FALSE, victim, NULL, NULL,
+      TO_ROOM, ANSI_CYAN);
+  } else
+    caster->nothingHappens();
+
+  return TRUE;
+}
+
+int detectShadow(TBeing* caster, TBeing* victim, int level, short bKnown) {
+  affectedData aff;
+
+  caster->reconcileHelp(victim, discArray[SPELL_DETECT_SHADOW]->alignMod);
+
+  if (caster->bSuccess(bKnown, SPELL_DETECT_SHADOW)) {
+    aff.type = SPELL_DETECT_SHADOW;
+    aff.duration = caster->durationModify(SPELL_DETECT_SHADOW,
+      level * 3 / 2 * Pulse::UPDATES_PER_MUDHOUR);
+    aff.modifier = 0;
+    aff.location = APPLY_NONE;
+    aff.bitvector = AFF_DETECT_INVISIBLE;
+
+    switch (critSuccess(caster, SPELL_DETECT_SHADOW)) {
+      case CRIT_S_DOUBLE:
+      case CRIT_S_TRIPLE:
+      case CRIT_S_KILL:
+        CS(SPELL_DETECT_SHADOW);
+        aff.duration *= 2;
+        break;
+      case CRIT_S_NONE:
+        break;
+    }
+    if (!victim->affectJoin(caster, &aff, AVG_DUR_NO, AVG_EFF_YES)) {
+      caster->nothingHappens();
+      return SPELL_FALSE;
+    }
+
+    act("$n's eyes briefly glow red.", FALSE, victim, 0, 0, TO_ROOM,
+      ANSI_RED_BOLD);
+    act("Your eyes sting.", FALSE, victim, 0, 0, TO_CHAR, ANSI_RED_BOLD);
+    return SPELL_SUCCESS;
+  } else {
+    caster->nothingHappens();
+    return SPELL_FAIL;
+  }
+}
+
+void detectShadow(TBeing* caster, TBeing* victim, TMagicItem* obj) {
+  detectShadow(caster, victim, obj->getMagicLevel(),
+    obj->getMagicLearnedness());
+}
+
+int detectShadow(TBeing* caster, TBeing* victim) {
+  taskDiffT diff;
+
+  if (!bPassShamanChecks(caster, SPELL_DETECT_SHADOW, victim))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_DETECT_SHADOW]->lag;
+  diff = discArray[SPELL_DETECT_SHADOW]->task;
+
+  start_cast(caster, victim, NULL, caster->roomp, SPELL_DETECT_SHADOW, diff, 1,
+    "", rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+int castDetectShadow(TBeing* caster, TBeing* victim) {
+  int ret, level;
+
+  level = caster->getSkillLevel(SPELL_DETECT_SHADOW);
+  int bKnown = caster->getSkillValue(SPELL_DETECT_SHADOW);
+
+  ret = detectShadow(caster, victim, level, bKnown);
+
+  if (ret == SPELL_SUCCESS) {}
+  return TRUE;
+}
+
+
+int djallasProtection(TBeing* caster, TBeing* victim, int level, short bKnown) {
+  affectedData aff, aff2, aff3, aff4;
+
+  aff.type = SPELL_DJALLA;
+  aff.level = level;
+  aff.duration = caster->durationModify(SPELL_DJALLA,
+    (3 + (level / 2)) * Pulse::UPDATES_PER_MUDHOUR);
+  aff.location = APPLY_IMMUNITY;
+  aff.modifier = IMMUNE_SUMMON;
+  aff.modifier2 = ((level * 2) / 3);
+  aff.bitvector = 0;
+
+  aff2.type = SPELL_DJALLA;
+  aff2.level = level;
+  aff2.duration = aff.duration;
+  aff2.location = APPLY_IMMUNITY;
+  aff2.modifier = IMMUNE_POISON;
+  aff2.modifier2 = ((level * 2) / 3);
+  aff2.bitvector = 0;
+
+  aff3.type = SPELL_DJALLA;
+  aff3.level = level;
+  aff3.duration = aff.duration;
+  aff3.location = APPLY_IMMUNITY;
+  aff3.modifier = IMMUNE_DRAIN;
+  aff3.modifier2 = ((level * 2) / 3);
+  aff3.bitvector = 0;
+
+  aff4.type = SPELL_DJALLA;
+  aff4.level = level;
+  aff4.duration = aff.duration;
+  aff4.location = APPLY_IMMUNITY;
+  aff4.modifier = IMMUNE_ENERGY;
+  aff4.modifier2 = ((level * 2) / 3);
+  aff4.bitvector = 0;
+
+  if (caster->bSuccess(bKnown, SPELL_DJALLA)) {
+    switch (critSuccess(caster, SPELL_DJALLA)) {
+      case CRIT_S_DOUBLE:
+      case CRIT_S_TRIPLE:
+      case CRIT_S_KILL:
+        CS(SPELL_DJALLA);
+        aff.duration *= 2;
+        aff.modifier2 = (level * 2);
+        aff2.duration *= 2;
+        aff2.modifier2 = (level * 2);
+        aff3.duration *= 2;
+        aff3.modifier2 = (level * 2);
+        aff4.duration *= 2;
+        aff4.modifier2 = (level * 2);
+        act("$n becomes one with the spirits.", FALSE, victim, NULL, NULL,
+          TO_ROOM, ANSI_GREEN);
+        act("You have been greatly blessed with the protection of Djalla!",
+          FALSE, victim, NULL, NULL, TO_CHAR, ANSI_GREEN);
+        break;
+      case CRIT_S_NONE:
+        act("$n becomes one with the spirits.", FALSE, victim, NULL, NULL,
+          TO_ROOM, ANSI_GREEN);
+        act("You have been granted the protection of Djalla!", FALSE, victim,
+          NULL, NULL, TO_CHAR, ANSI_GREEN);
+        break;
+    }
+
+    if (caster != victim) {
+      aff.modifier2 /= 2;
+      aff2.modifier2 /= 2;
+      aff3.modifier2 /= 2;
+      aff4.modifier2 /= 2;
+    }
+
+    victim->affectJoin(caster, &aff, AVG_DUR_NO, AVG_EFF_YES);
+    victim->affectJoin(caster, &aff2, AVG_DUR_NO, AVG_EFF_YES);
+    victim->affectJoin(caster, &aff3, AVG_DUR_NO, AVG_EFF_YES);
+    victim->affectJoin(caster, &aff4, AVG_DUR_NO, AVG_EFF_YES);
+    caster->reconcileHelp(victim, discArray[SPELL_DJALLA]->alignMod);
+    return SPELL_SUCCESS;
+  } else {
+    caster->nothingHappens();
+    return SPELL_FAIL;
+  }
+}
+void djallasProtection(TBeing* caster, TBeing* victim, TMagicItem* obj) {
+  djallasProtection(caster, victim, obj->getMagicLevel(),
+    obj->getMagicLearnedness());
+}
+
+int djallasProtection(TBeing* caster, TBeing* victim) {
+  if (!bPassShamanChecks(caster, SPELL_DJALLA, victim))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_DJALLA]->lag;
+  taskDiffT diff = discArray[SPELL_DJALLA]->task;
+
+  start_cast(caster, victim, NULL, caster->roomp, SPELL_DJALLA, diff, 1, "",
+    rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+int castDjallasProtection(TBeing* caster, TBeing* victim) {
+  int level = caster->getSkillLevel(SPELL_DJALLA);
+  int bKnown = caster->getSkillValue(SPELL_DJALLA);
+
+  int ret = djallasProtection(caster, victim, level, bKnown);
+  if (ret == SPELL_SUCCESS) {
+  } else {
+  }
+  return TRUE;
+}
+
+int legbasGuidance(TBeing* caster, TBeing* victim, int level, short bKnown) {
+  affectedData aff, aff2, aff3, aff4;
+
+  aff.type = SPELL_LEGBA;
+  aff.level = level;
+  aff.duration = caster->durationModify(SPELL_LEGBA,
+    (3 + (level / 2)) * Pulse::UPDATES_PER_MUDHOUR);
+  aff.location = APPLY_IMMUNITY;
+  aff.modifier = IMMUNE_BLEED;
+  aff.modifier2 = ((level * 2) / 3);
+  aff.bitvector = 0;
+
+  aff2.type = SPELL_LEGBA;
+  aff2.level = level;
+  aff2.duration = aff.duration;
+  aff2.location = APPLY_IMMUNITY;
+  aff2.modifier = IMMUNE_EARTH;
+  aff2.modifier2 = ((level * 2) / 3);
+  aff2.bitvector = 0;
+
+  aff3.type = SPELL_LEGBA;
+  aff3.level = level;
+  aff3.duration = aff.duration;
+  aff3.location = APPLY_IMMUNITY;
+  aff3.modifier = IMMUNE_CHARM;
+  aff3.modifier2 = ((level * 2) / 3);
+  aff3.bitvector = 0;
+
+  aff4.type = SPELL_LEGBA;
+  aff4.level = level;
+  aff4.duration = aff.duration;
+  aff4.location = APPLY_IMMUNITY;
+  aff4.modifier = IMMUNE_SLEEP;
+  aff4.modifier2 = ((level * 2) / 3);
+  aff4.bitvector = 0;
+
+  if (caster->bSuccess(bKnown, SPELL_LEGBA)) {
+    switch (critSuccess(caster, SPELL_LEGBA)) {
+      case CRIT_S_DOUBLE:
+      case CRIT_S_TRIPLE:
+      case CRIT_S_KILL:
+        CS(SPELL_LEGBA);
+        aff.duration *= 2;
+        aff.modifier2 = (level * 2);
+        aff2.duration *= 2;
+        aff2.modifier2 = (level * 2);
+        aff3.duration *= 2;
+        aff3.modifier2 = (level * 2);
+        aff4.duration *= 2;
+        aff4.modifier2 = (level * 2);
+        act("$n becomes one with the spirits.", FALSE, victim, NULL, NULL,
+          TO_ROOM, ANSI_GREEN);
+        act("You have been greatly blessed with the protection of Legba!",
+          FALSE, victim, NULL, NULL, TO_CHAR, ANSI_GREEN);
+        break;
+      case CRIT_S_NONE:
+        act("$n becomes one with the spirits.", FALSE, victim, NULL, NULL,
+          TO_ROOM, ANSI_GREEN);
+        act("You have been granted the protection of Legba!", FALSE, victim,
+          NULL, NULL, TO_CHAR, ANSI_GREEN);
+        break;
+    }
+
+    if (caster != victim) {
+      aff.modifier2 /= 2;
+      aff2.modifier2 /= 2;
+      aff3.modifier2 /= 2;
+      aff4.modifier2 /= 2;
+    }
+
+    victim->affectJoin(caster, &aff, AVG_DUR_NO, AVG_EFF_YES);
+    victim->affectJoin(caster, &aff2, AVG_DUR_NO, AVG_EFF_YES);
+    victim->affectJoin(caster, &aff3, AVG_DUR_NO, AVG_EFF_YES);
+    victim->affectJoin(caster, &aff4, AVG_DUR_NO, AVG_EFF_YES);
+    caster->reconcileHelp(victim, discArray[SPELL_LEGBA]->alignMod);
+    return SPELL_SUCCESS;
+  } else {
+    caster->nothingHappens();
+    return SPELL_FAIL;
+  }
+}
+void legbasGuidance(TBeing* caster, TBeing* victim, TMagicItem* obj) {
+  legbasGuidance(caster, victim, obj->getMagicLevel(),
+    obj->getMagicLearnedness());
+}
+
+int legbasGuidance(TBeing* caster, TBeing* victim) {
+  if (!bPassShamanChecks(caster, SPELL_LEGBA, victim))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_LEGBA]->lag;
+  taskDiffT diff = discArray[SPELL_LEGBA]->task;
+
+  start_cast(caster, victim, NULL, caster->roomp, SPELL_LEGBA, diff, 1, "",
+    rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+int castLegbasGuidance(TBeing* caster, TBeing* victim) {
+  int level = caster->getSkillLevel(SPELL_LEGBA);
+  int bKnown = caster->getSkillValue(SPELL_LEGBA);
+
+  int ret = legbasGuidance(caster, victim, level, bKnown);
+  if (ret == SPELL_SUCCESS) {
+  } else {
+  }
+  return TRUE;
+}
+
+int chaseSpirits(TBeing* caster, TObj* obj, int, short bKnown) {
+  int i;
+
+  if (caster->bSuccess(bKnown, SPELL_CHASE_SPIRIT)) {
+    for (i = 0; i < MAX_OBJ_AFFECT; i++) {
+      if ((obj->affected[i].location != APPLY_NONE) &&
+          (obj->affected[i].location != APPLY_LIGHT) &&
+          (obj->affected[i].location != APPLY_NOISE) &&
+          (obj->affected[i].location != APPLY_HIT) &&
+          (obj->affected[i].location != APPLY_CHAR_WEIGHT) &&
+          (obj->affected[i].location != APPLY_CHAR_HEIGHT) &&
+          (obj->affected[i].location != APPLY_MOVE) &&
+          (obj->affected[i].location != APPLY_ARMOR)) {
+        obj->affected[i].location = APPLY_NONE;
+        obj->affected[i].modifier = 0;
+        obj->affected[i].modifier2 = 0;
+        obj->affected[i].bitvector = 0;
+      }
+    }
+    obj->remObjStat(ITEM_MAGIC);
+
+    return SPELL_SUCCESS;
+  } else {
+    caster->nothingHappens();
+    return SPELL_FAIL;
+  }
+}
+
+void chaseSpirits(TBeing* caster, TObj* tar_obj, TMagicItem* obj) {
+  int ret = 0;
+  int level = obj->getMagicLevel();
+
+  ret = chaseSpirits(caster, tar_obj, level, obj->getMagicLearnedness());
+  if (IS_SET(ret, SPELL_SUCCESS)) {
+    act("$p chases the spirits from $n...", FALSE, tar_obj, obj, NULL, TO_ROOM);
+  }
+}
+
+int chaseSpirits(TBeing* caster, TObj* obj) {
+  taskDiffT diff;
+
+  if (!bPassShamanChecks(caster, SPELL_CHASE_SPIRIT, obj))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_CHASE_SPIRIT]->lag;
+  diff = discArray[SPELL_CHASE_SPIRIT]->task;
+
+  start_cast(caster, NULL, obj, caster->roomp, SPELL_CHASE_SPIRIT, diff, 2, "",
+    rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+int castChaseSpirits(TBeing* caster, TObj* obj) {
+  int ret = 0, level;
+
+  level = caster->getSkillLevel(SPELL_CHASE_SPIRIT);
+
+  ret =
+    chaseSpirits(caster, obj, level, caster->getSkillValue(SPELL_CHASE_SPIRIT));
+
+  if (IS_SET(ret, SPELL_SUCCESS)) {
+    act("You chase away the evil spirits within $N...", FALSE, caster, NULL,
+      obj, TO_CHAR);
+    act("$n chases away the evil spirits within $N...", FALSE, caster, NULL,
+      obj, TO_ROOM);
+  }
+  return ret;
+}
+
+int chaseSpirits(TBeing* caster, TBeing* victim, int, short bKnown) {
+  caster->reconcileHurt(victim, discArray[SPELL_CHASE_SPIRIT]->alignMod);
+
+  if (caster->bSuccess(bKnown, SPELL_CHASE_SPIRIT)) {
+    return SPELL_SUCCESS;
+  } else {
+    caster->nothingHappens();
+    return SPELL_FAIL;
+  }
+}
+
+int chaseSpirits(TBeing* caster, TBeing* victim, TMagicItem* obj) {
+  mud_assert(caster != NULL, "chaseSpirits(): no caster");
+  mud_assert(victim != NULL, "chaseSpirits(): no victim");
+
+  int level = obj->getMagicLevel();
+
+  int ret = chaseSpirits(caster, victim, level, obj->getMagicLearnedness());
+  if (IS_SET(ret, SPELL_SUCCESS)) {
+    act("$p chases away the evil spirits within you...", FALSE, victim, obj,
+      NULL, TO_CHAR);
+    act("$p chases away the evil spirits within $n...", FALSE, victim, obj,
+      NULL, TO_ROOM);
+    int rc = genericChaseSpirits(caster, victim, level, caster->isImmortal());
+    if (IS_SET_DELETE(rc, DELETE_VICT))
+      return DELETE_VICT;
+  }
+  return 0;
+}
+
+int chaseSpirits(TBeing* caster, TBeing* victim) {
+  if (!bPassShamanChecks(caster, SPELL_CHASE_SPIRIT, victim))
+    return FALSE;
+
+  lag_t rounds = discArray[SPELL_CHASE_SPIRIT]->lag;
+  taskDiffT diff = discArray[SPELL_CHASE_SPIRIT]->task;
+
+  start_cast(caster, victim, NULL, caster->roomp, SPELL_CHASE_SPIRIT, diff, 1,
+    "", rounds, caster->in_room, 0, 0, TRUE, 0);
+  return TRUE;
+}
+
+int castChaseSpirits(TBeing* caster, TBeing* victim) {
+  mud_assert(caster != NULL, "castChaseSpirits(): no caster");
+  mud_assert(victim != NULL, "castChaseSpirits(): no victim");
+
+  int level = caster->getSkillLevel(SPELL_CHASE_SPIRIT);
+  if (caster->isNotPowerful(victim, level, SPELL_CHASE_SPIRIT, SILENT_NO)) {
+    return 0;
+  }
+
+  int ret = chaseSpirits(caster, victim, level,
+    caster->getSkillValue(SPELL_CHASE_SPIRIT));
+
+  if (IS_SET(ret, SPELL_SUCCESS)) {
+    if (caster != victim) {
+      act("You chase away the evil spirits within $N...", FALSE, caster, NULL,
+        victim, TO_CHAR);
+      act("$n chases away the evil spirits within you...", FALSE, caster, NULL,
+        victim, TO_VICT);
+      act("$n chases away the evil spirits within $N...", FALSE, caster, NULL,
+        victim, TO_NOTVICT);
+    } else {
+      act("The loa chases away the evil within you...", FALSE, caster, NULL, 0,
+        TO_CHAR);
+      act("$n's face glows with much relief.", FALSE, caster, NULL, 0, TO_ROOM);
+    }
+    int rc = genericChaseSpirits(caster, victim, level, caster->isImmortal());
+    if (IS_SET_DELETE(rc, DELETE_VICT))
+      return DELETE_VICT;
+  }
+  return TRUE;
+}
+
+// returns DELETE_VICT (vict)
+int genericChaseSpirits(TBeing* caster, TBeing* victim, int,
+  immortalTypeT immortal, safeTypeT safe) {
+  mud_assert(victim != NULL, "genericChaseSpirits(): no victim");
+
+  TMonster* tvm = dynamic_cast<TMonster*>(victim);
+  spellNumT spell;
+  int rc;
+
+  struct chaseStruct {
+      spellNumT spell;
+      bool aggressive_act;
+      bool needs_saving_throw;
+      bool death_time_only;  // this is primarily for prayers
+  };
+
+  chaseStruct chaseArray[] = {
+    // air disc
+    {SPELL_FEATHERY_DESCENT, false, true, false},
+    {SPELL_FLY, true, true, false},
+    {SPELL_ANTIGRAVITY, true, true, false},
+    {SPELL_LEVITATE, true, true, false},
+    {SPELL_FALCON_WINGS, true, true, false},
+    {SPELL_PROTECTION_FROM_AIR, true, true, false},
+    // alchemy
+    {SPELL_DETECT_MAGIC, false, true, false},
+    // earth
+    {SPELL_STONE_SKIN, false, true, false},
+    {SPELL_TRAIL_SEEK, false, true, false},
+    {SPELL_PROTECTION_FROM_EARTH, true, true, false},
+    // fire
+    {SPELL_FAERIE_FIRE, false, false, false},
+    {SPELL_FLAMING_FLESH, false, true, false},
+    {SPELL_INFRAVISION, false, true, false},
+    {SPELL_PROTECTION_FROM_FIRE, true, true, false},
+    // sorcery
+    {SPELL_SORCERERS_GLOBE, true, true, false},
+    {SPELL_BIND, false, false, false},
+    {SPELL_PROTECTION_FROM_ENERGY, true, true, false},
+    // spirit
+    {SPELL_SILENCE, false, true, false},
+    {SPELL_ENSORCER, false, false, false},
+    {SPELL_INVISIBILITY, false, true, false},
+    {SPELL_STEALTH, false, true, false},
+    {SPELL_ACCELERATE, true, true, false},
+    {SPELL_HASTE, true, true, false},
+    {SPELL_CALM, false, true, false},
+    {SPELL_SENSE_LIFE, false, true, false},
+    {SPELL_DETECT_INVISIBLE, false, true, false},
+    {SPELL_TRUE_SIGHT, false, true, false},
+    {SPELL_FEAR, false, false, false},
+    {SPELL_SLUMBER, false, false, false},
+    // water
+    {SPELL_ICY_GRIP, false, false, false},
+    {SPELL_GILLS_OF_FLESH, true, true, false},
+    {SPELL_PROTECTION_FROM_WATER, true, true, false},
+    {SPELL_PLASMA_MIRROR, true, true, false},
+    {SPELL_GARMULS_TAIL, false, true, false},
+
+    // cleric prayers - these should be death-time only stuff
+    {SPELL_SANCTUARY, true, true, true},
+    {SPELL_ARMOR, true, true, true},
+    {SPELL_ARMOR_DEIKHAN, true, true, true},
+    {SPELL_BLESS, true, true, true},
+    {SPELL_BLESS_DEIKHAN, true, true, true},
+    {SPELL_BLINDNESS, false, false, true},
+    {SPELL_PARALYZE, false, false, true},
+    {SPELL_POISON, false, false, true},
+    {SPELL_POISON_DEIKHAN, false, false, true},
+    {SPELL_CURSE, false, false, true},
+    {SPELL_CURSE_DEIKHAN, false, false, true},
+    // shaman stuff
+    {SPELL_STUPIDITY, true, true, false},
+    {SPELL_CELERITE, true, true, false},
+    {SPELL_LEGBA, true, true, false},
+    {SPELL_DJALLA, true, true, false},
+    {SPELL_SENSE_LIFE_SHAMAN, true, true, false},
+    {SPELL_DETECT_SHADOW, true, true, false},
+    {SPELL_INTIMIDATE, true, true, false},
+    {SPELL_CHEVAL, true, true, false},
+    {SPELL_HYPNOSIS, true, true, false},
+    {SPELL_CLARITY, true, true, false},
+    {SPELL_AQUALUNG, true, true, false},
+    {SPELL_THORNFLESH, true, true, false},
+    {SPELL_SHIELD_OF_MISTS, true, true, false},
+    {SPELL_CONTROL_UNDEAD, true, true, false},
+    {SPELL_RESURRECTION, true, true, false},
+    {SPELL_DANCING_BONES, true, true, false},
+    {SPELL_VOODOO, true, true, false},
+
+#if 0
+    // these effects are on mobs
+    // death-time-only is silly to check for
+    { SPELL_STICKS_TO_SNAKES, false, false, true },
+    { SPELL_LIVING_VINES, false, false, true },
+    { SPELL_PLAGUE_LOCUSTS, false, false, true },
+#endif
+
+#if 0
+    // not yet implemented
+    { SPELL_DETECT_POISON, false, true, false },
+    { SPELL_DETECT_POISON_DEIKHAN, false, true, false },
+#endif
+
+#if 0
+    // skills that should be usable again if they die
+    // these use to have a check for ARENA-death
+    // not sure how to do this in new setup, so commented out for time being
+    { SKILL_TRANSFIX, false, false, true },
+    { SKILL_CHI, false, false, true },
+    { SKILL_DOORBASH, false, false, true },
+    { SKILL_TRANSFORM_LIMB, false, false, true },
+    { SKILL_BARKSKIN, false, false, true },
+    { SKILL_TRACK, false, false, true },
+    { SKILL_CONCEALMENT, false, false, true },
+    { SKILL_FORAGE, false, false, true },
+    { SKILL_SEEKWATER, false, false, true },
+    { SKILL_ENCAMP, false, false, true },
+    { SKILL_DIVINATION, false, false, true },
+    { SKILL_SPY, false, false, true },
+    { SKILL_DISGUISE, false, false, true },
+    { SKILL_BERSERK, false, false, true },
+    { SKILL_DEATHSTROKE, false, false, true },
+    { SKILL_DOORBASH, false, false, true },
+    { SKILL_QUIV_PALM, false, false, true },
+#endif
+
+    {TYPE_UNDEFINED, false, false, false}  // this is final terminator
+    // spell, aggressive, saving throw, death_time_only
+  };
+
+  int iter;
+  for (iter = 0; chaseArray[iter].spell != TYPE_UNDEFINED; iter++) {
+    spell = chaseArray[iter].spell;
+
+    // check if they have the spell
+    // should decay if !caster (death-time) or if set to decay all the time
+    if ((!caster || !chaseArray[iter].death_time_only) &&
+        victim->affectedBySpell(spell)) {
+      // immortals should always succeed
+      // make a save otherwise
+      // there is assumption that !caster (death-time) will have immortal=true
+      if (immortal || !chaseArray[iter].needs_saving_throw ||
+          !victim->isLucky(caster->spellLuckModifier(SPELL_CHASE_SPIRIT))) {
+        rc = victim->spellWearOff(spell, safe);
+        if (IS_SET_DELETE(rc, DELETE_THIS))
+          return DELETE_VICT;
+        victim->affectFrom(spell);
+      }
+      // aggressive Act
+      if (caster && !victim->fight() && tvm) {
+        caster->setCharFighting(victim);
+        caster->setVictFighting(victim);
+      }
+    }
+  }
+
+  if (caster && victim->isAffected(AFF_SANCTUARY)) {
+    if (immortal ||
+        !victim->isLucky(caster->spellLuckModifier(SPELL_CHASE_SPIRIT))) {
+      REMOVE_BIT(victim->specials.affectedBy, AFF_SANCTUARY);
+      victim->sendTo(
+        "You feel more vulnerable as your white aura slowly fades.\n\r");
+      act("The white glow around $n's body fades.", FALSE, victim, NULL, NULL,
+        TO_ROOM);
+    }
+    // aggressive Act
+    if (caster && !victim->fight() && tvm) {
+      caster->setCharFighting(victim);
+      caster->setVictFighting(victim);
+    }
+  }
+  return FALSE;
 }
