+++ code/code/disc/disc_warrior_brawling.h
@@ -22,3 +22,5 @@ class CDBrawling : public CDiscipline {
 
   private:
 };
+
+int berserk(TBeing*);
