+++ code/code/misc/spelltask.cc
@@ -62,6 +62,7 @@
 #include "disc_shaman_spider.h"
 #include "disc_shaman_skunk.h"
 #include "disc_shaman_control.h"
+#include "disc_shaman.h"
 #include "disc_ranger_nature.h"
 #include "disc_cleric_cures.h"
 #include "disc_mage_earth.h"
@@ -71,7 +72,7 @@
 #include "disc_mage_water.h"
 #include "disc_cleric_aegis.h"
 #include "disc_cleric_wrath.h"
-#include "disc_shaman.h"
+#include "disc_shaman_alchemy.h"
 #include "disc_cleric_afflictions.h"
 #include "disc_cleric_cures.h"
 #include "disc_cleric_hand_of_god.h"
